added dependency_changes_lib
