--- timeline-dep-graph/src/app/timeline-dep-graph/timeline/dependencyChanges.ts
@@ -1,15 +1,0 @@
-import { Task } from '../Task';
-
-export interface DependecyChanges {
-  add: Task[];
-  remove: Task[];
-  update: Task[];
-}
-
-export function createDependecyChanges(): DependecyChanges {
-  return {
-    add: [],
-    remove: [],
-    update: [],
-  };
-}
