+++ timeline-dep-graph/src/app/timeline-dep-graph/timeline/dependency_changes_lib..ts
@@ -1,0 +1,55 @@
+import { Task } from '../Task';
+import { equalsTask, TaskId } from './../Task';
+
+interface TaskChangesHolder {
+  prev?: Task;
+  curr?: Task;
+}
+
+export interface DependecyChanges {
+  add: Task[];
+  remove: Task[];
+  update: Task[];
+}
+
+/**
+ * @param prev The previous state of tasks array.
+ * @param curr The current state of tasks array.
+ * @return     The changes between the previous state and the current state.
+ */
+export function getDependecyChanges(prev: Task[], curr: Task[])
+  : DependecyChanges {
+  const map = new Map<TaskId, TaskChangesHolder>();
+
+  for (const task of prev) {
+    map.set(task.id, { prev: task });
+  }
+
+  for (const task of curr) {
+    if (map.has(task.id)) {
+      map.get(task.id).curr = task;
+    } else {
+      map.set(task.id, { curr: task });
+    }
+  }
+
+  const updatedTasks: DependecyChanges = {
+    add: [],
+    remove: [],
+    update: [],
+  };
+  for (const val of map.values()) {
+    const prevTask = val.prev;
+    const currTask = val.curr;
+
+    if (!currTask) {
+      updatedTasks.remove.push(prevTask);
+    } else if (!prevTask) {
+      updatedTasks.add.push(currTask);
+    } else if (!equalsTask(currTask, prevTask)) {
+      updatedTasks.update.push(currTask);
+    }
+  }
+
+  return updatedTasks;
+}
